# Conflicts:
#	app/Domains/Post/Content/Nodes/Callout.php
#	app/Domains/Post/Content/Nodes/CodeBlock.php
#	app/Domains/Post/PostRepository.php

+++ app/Domains/Post/PostRepository.php
@@ -4,6 +4,7 @@ namespace App\Domains\Post;
 
 use App\Data\Enums\PostStatusEnum;
 use App\Domains\Language\LanguageRepository;
+use App\Domains\Post\Content\PostContentService;
 use App\Domains\Post\Events\PostCreatedEvent;
 use App\Domains\Post\Events\PostDeletedEvent;
 use App\Domains\Post\Events\PostUpdatedEvent;
@@ -83,8 +84,8 @@ class PostRepository
                             COALESCE(posts.published_at, posts.created_at) < ?
                         ',
                         [
-                            Carbon::createFromTimestamp($startTimestamp)->toDateTimeString(),
-                            Carbon::createFromTimestamp($endTimestamp)->toDateTimeString(),
+                            Carbon::createFromTimestamp((int) $startTimestamp)->toDateTimeString(),
+                            Carbon::createFromTimestamp((int) $endTimestamp)->toDateTimeString(),
                         ]
                     );
             })
@@ -263,7 +264,10 @@ class PostRepository
          *
          * #ref https://github.com/laravel/framework/issues/21449
          */
-        return Post::find($post->id);
+        /** @var Post $post */
+        $post = Post::find($post->id);
+
+        return $post;
     }
 
     /**
@@ -306,7 +310,7 @@ class PostRepository
         return $post;
     }
 
-    public static function deletePost(Post $post)
+    public static function deletePost(Post $post) : void
     {
         $post->variants->map(fn ($variant) => self::deletePostVariant($post, $variant->language_id));
         $post->delete();
@@ -417,14 +421,34 @@ class PostRepository
             ->first();
     }
 
-    public static function deletePostVariant(Post $post, int $languageId)
+    public static function deletePostVariant(Post $post, int $languageId) : void
     {
         $variant = PostVariant::where('language_id', $languageId)
             ->where('post_id', $post->id)
             ->first();
 
-        $variant->delete();
+        if (!$variant)
+            return;
 
+        $variant->delete();
         PostVariantDeletedEvent::dispatch($variant);
+    }
+
+    public static function updateVariantHtml(PostVariant $variant): void
+    {
+        if (!$variant->content) {
+            return;
+        }
+
+        $post = $variant->post;
+        if (!$post) return;
+
+        $blog = $post->blog;
+        if (!$blog) return;
+
+        $html = PostContentService::getHtml($variant->content, $blog);
+
+        $variant->content_html = $html;
+        $variant->save();
     }
 }
